# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Australian Economic Dashboard API",
    "description": "API for Australian economic data — energy, housing, and prices",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/energy/overview": {
      "get": {
        "tags": [
          "crate::routes::energy"
        ],
        "operationId": "overview",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "description": "Region code (default AU)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnergyOverviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/housing/overview": {
      "get": {
        "tags": [
          "crate::routes::housing"
        ],
        "operationId": "overview",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "description": "Region code (default AU)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HousingOverviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/oil/import-sources": {
      "get": {
        "tags": [
          "crate::routes::oil"
        ],
        "operationId": "import_sources",
        "parameters": [
          {
            "name": "hs_code",
            "in": "query",
            "description": "HS commodity code (default 2709)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Year filter (e.g. 2024)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OilImportSourcesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/oil/overview": {
      "get": {
        "tags": [
          "crate::routes::oil"
        ],
        "operationId": "overview",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "description": "Region code (default AU)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OilOverviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/oil/timeseries": {
      "get": {
        "tags": [
          "crate::routes::oil"
        ],
        "operationId": "timeseries",
        "parameters": [
          {
            "name": "series_id",
            "in": "query",
            "description": "Oil series ID (e.g. oil.production.kbd)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "description": "Region code (default AU)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OilTimeSeriesResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnergyOverviewResponse": {
        "type": "object",
        "required": [
          "region",
          "methodSummary",
          "sourceRefs",
          "sourceMixViews",
          "panels",
          "freshness"
        ],
        "properties": {
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "methodSummary": {
            "type": "string"
          },
          "panels": {
            "$ref": "#/components/schemas/EnergyPanels"
          },
          "region": {
            "type": "string"
          },
          "sourceMixViews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceMixViewDto"
            }
          },
          "sourceRefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          }
        }
      },
      "EnergyPanels": {
        "type": "object",
        "properties": {
          "benchmark": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PanelBenchmark"
              }
            ]
          },
          "cpiElectricity": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PanelCpi"
              }
            ]
          },
          "liveWholesale": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PanelWholesale"
              }
            ]
          },
          "retailAverage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PanelRetail"
              }
            ]
          }
        }
      },
...
      "FreshnessInfo": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "HousingMetric": {
        "type": "object",
        "required": [
          "seriesId",
          "date",
          "value"
        ],
        "properties": {
          "date": {
            "type": "string"
          },
          "seriesId": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "HousingOverviewResponse": {
        "type": "object",
        "required": [
          "region",
          "requiredSeriesIds",
          "missingSeriesIds",
          "metrics"
        ],
        "properties": {
          "metrics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HousingMetric"
            }
          },
          "missingSeriesIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "region": {
            "type": "string"
          },
          "requiredSeriesIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "OilImportSource": {
        "type": "object",
        "required": [
          "countryCode",
          "countryName",
          "valueUsd",
          "sharePct",
          "period"
        ],
        "properties": {
          "countryCode": {
            "type": "string"
          },
          "countryName": {
            "type": "string"
          },
          "period": {
            "type": "string"
          },
          "sharePct": {
            "type": "number",
            "format": "double"
          },
          "valueUsd": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "OilImportSourcesResponse": {
        "type": "object",
        "required": [
          "period",
          "totalValueUsd",
          "sources",
          "sourceRefs",
          "freshness"
        ],
        "properties": {
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "period": {
            "type": "string"
          },
          "sourceRefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OilImportSource"
            }
          },
          "totalValueUsd": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "OilMetricPoint": {
        "type": "object",
        "required": [
          "period",
          "valueKbd",
          "countryCode"
        ],
        "properties": {
          "countryCode": {
            "type": "string"
          },
          "period": {
            "type": "string"
          },
          "valueKbd": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "OilOverviewResponse": {
        "type": "object",
        "required": [
          "region",
          "sourceRefs",
          "freshness"
        ],
        "properties": {
          "consumption": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OilMetricPoint"
              }
            ]
          },
          "exports": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OilMetricPoint"
              }
            ]
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "imports": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OilMetricPoint"
              }
            ]
          },
          "production": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OilMetricPoint"
              }
            ]
          },
          "region": {
            "type": "string"
          },
          "sourceRefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          }
        }
      },
      "OilTimeSeriesPoint": {
        "type": "object",
        "required": [
          "period",
          "valueKbd"
        ],
        "properties": {
          "period": {
            "type": "string"
          },
          "valueKbd": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "OilTimeSeriesResponse": {
        "type": "object",
        "required": [
          "seriesId",
          "region",
          "points",
          "sourceRefs",
          "freshness"
        ],
        "properties": {
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OilTimeSeriesPoint"
            }
          },
          "region": {
            "type": "string"
          },
          "seriesId": {
            "type": "string"
          },
          "sourceRefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          }
        }
      },
      "PanelBenchmark": {
        "type": "object",
        "required": [
          "dmoAnnualBillAud"
        ],
        "properties": {
          "dmoAnnualBillAud": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PanelCpi": {
        "type": "object",
        "required": [
          "indexValue",
          "period"
        ],
        "properties": {
          "indexValue": {
            "type": "number",
            "format": "double"
          },
          "period": {
            "type": "string"
          }
        }
      },
      "PanelRetail": {
        "type": "object",
        "required": [
          "annualBillAudMean",
          "annualBillAudMedian"
        ],
        "properties": {
          "annualBillAudMean": {
            "type": "number",
            "format": "double"
          },
          "annualBillAudMedian": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PanelWholesale": {
        "type": "object",
        "required": [
          "valueAudMwh",
          "valueCKwh"
        ],
        "properties": {
          "valueAudMwh": {
            "type": "number",
            "format": "double"
          },
          "valueCKwh": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "SourceMixRowDto": {
        "type": "object",
        "required": [
          "sourceKey",
          "label",
          "sharePct"
        ],
        "properties": {
          "label": {
            "type": "string"
          },
          "sharePct": {
            "type": "number",
            "format": "double"
          },
          "sourceKey": {
            "type": "string"
          }
        }
      },
      "SourceMixViewDto": {
        "type": "object",
        "required": [
          "viewId",
          "title",
          "coverageLabel",
          "sourceRefs",
          "rows"
        ],
        "properties": {
          "coverageLabel": {
            "type": "string"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceMixRowDto"
            }
          },
          "sourceRefs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            }
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "viewId": {
            "type": "string"
          }
        }
      },
      "SourceRef": {
        "type": "object",
        "required": [
          "sourceId",
          "name",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
    }
  }
}